Support ID definition inside embedded struct
There is a case when the Object ID tag is defined inside of the Embedded
struct e.g. id in base entity object. To support this case, I check
whether the object ID of nested diffs is available or not. It will be
still checked if the ID tag is defined in both main and embedded struct.