Fix compare map with array value
There is a bug when we compare the maps with array value. Turns out,
instead comparing the filtered value, the library still comparing the
unfiltered value, which in this case, it will defined as an interface{}.
So, I fix it by changing the value to be compared and add a unit test
for this case.