Refactor get object ID function
There is a case when we put the Object ID tag in the embedded struct.
Since previously, if there any nested or embedded struct changes, we get
the Object ID from the first diff, and pass it to all object ID.

This way would not work but
if there is no change in that embedded struct. So I fix it by creating
separated getObjectID function that scan all fields in the struct.