Refactor Compare function (#15)
I start to feel the pain when trying to enhance the code to cover the edge cases. So, I decided to refactor the Compare function by separating each of the kind comparators. I also move the Diff struct from libra package to diff package, to avoid cyclic between the library.